Set utop and `cs3110 compile` to use the -safe-string flag
Ocaml 4.02 introduced the type `bytes` to enforce a separation between
mutable strings (the new `bytes` type) and immutable strings (the`string`
type). However, for backwards compatibility purposes the -unsafe-string
flag is the current default, meaning the two are interchangeable strings
become a mutable data structure (not what we generally want in 3110), and
utop says strings are of type `bytes` which is confusing.

This change simply adds the -safe-string flag (which the OCaml
documentation recommends all new code has) to invocations of `utop` and
`cs3110 compile`.