Harness: Fix regular expression to capture test results.
Because the capture after "characters" was just .* and left at the
default greedy operation, if the error message had a : anywhere else in
the line then the capture group would be incorrect, and would only
capture the text after the : character and thus miss the test name. This
patch makes the regex more explicit here.